+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+### Added
+
+- DelayUs and DelayMs trait implementations for timer
+
 ## [0.2.1]
 
 ### Added